refactor: remove ternary operator in component
