+++ src/components/UI/itemProperty/itemProperty.tsx
@@ -6,14 +6,12 @@ export type ItemPropertyProps = {
 export function ItemProperty(props: ItemPropertyProps) {
   return (
     <>
-      {props.value ? (
-        <p>
-          <span>{props.text}</span>
-          <span>{props.value}</span>
-        </p>
-      ) : (
-        <></>
-      )}
+      props.value && (
+      <p>
+        <span>{props.text}</span>
+        <span>{props.value}</span>
+      </p>
+      )
     </>
   );
 }
