+++ src/components/UI/itemRatingProperty/itemRatingProperty.tsx
@@ -8,14 +8,12 @@ export type ItemRatingPropertyProps = {
 export function ItemRatingProperty(props: ItemRatingPropertyProps) {
   return (
     <>
-      {props.rating ? (
+      {props.rating && (
         <p>
           <span>{props.children}</span>
           <span>{props.rating}</span>
           <span>Voted: ({props.voted})</span>
         </p>
-      ) : (
-        <></>
       )}
     </>
   );
